requirements.txt
opencv-python>=4.8.0
PyMuPDF>=1.23.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0
reportlab>=4.0.0 